fix grumbles: move client error to own mod
